VBOS and VAOS ERMAHGERD
https://pbs.twimg.com/profile_images/417190013160685568/AOe97q30.jpeg

+++ Halite/Classes/GameLogic/hlt.h
@@ -35,7 +35,7 @@ namespace hlt
 
 	struct Color
 	{
-		GLubyte r, g, b;
+		float r, g, b;
 	};
 
 	struct Site
